requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
python-multipart==0.0.6
python-jose==3.3.0
cryptography==41.0.7
passlib==1.7.4
redis==5.0.1
pillow==10.1.0
websockets==12.0
pydantic-settings==2.1.0Pillow==10.1.0
